rust: kernel: add support for bits/genmask macros
These macros are converted from their kernel equivalent.

Note by Remo: During a rebase, this file seems to have been added to the
mainline kernel. The macros it contains all had different names though.
So I just kept the content from both sides. Subsequent patches that rely
on the "outdated" version of this file should thus still work. But, this
is not clean at all and should be looked at in detail before anything
that depends on it is considered for upstreaming.